Change `ChangeContext` implementation to force initialization
Roslyn does not complain when declaring a constructor for a `struct`
that contains a sole parameter with a default value, but the constructor
is never invoked with the default value if called without parameters
(the default `struct` constructor is used, instead).

Make the old `IDisposable` `ChangeContext` a private inner class and
have the new `ChangeContext` return an opaque instance of it as an
`IDisposable` to prevent misuse.